Optimize data types: byte-back all enums, narrow fields, fix size mismatch
- Add explicit `: byte` backing to all ~170 remaining enums across the
  entire codebase (Application, Contracts, Core, Domain, Infrastructure,
  ProviderSdk, Storage, Ui.Services, Wpf). Each enum field drops from
  4 bytes to 1 byte in memory and serialized output.
- Narrow IntegrityEvent.ErrorCode from int? to ushort? (codes 1001-1007)
- Narrow OptionContractSpec.Multiplier from int to ushort (typically 100)
- Narrow MarketDepthCollector.MaxDepth constant from int to ushort
- Fix Alpaca wire message type mismatch: AlpacaTradeMessage.Size,
  AlpacaQuoteMessage.BidSize/AskSize changed from int to long to match
  domain model types and prevent silent truncation on large volumes
- Remove tautological Multiplier <= 0 check (ushort can't be negative)

+++ src/MarketDataCollector.Contracts/Domain/Models/IntegrityEvent.cs
@@ -11,7 +11,7 @@ public sealed record IntegrityEvent(
     string Symbol,
     IntegritySeverity Severity,
     string Description,
-    int? ErrorCode,
+    ushort? ErrorCode,
     long SequenceNumber,
     string? StreamId = null,
     string? Venue = null
